+++ src/main/java/com/foodease/myapp/service/dto/response/OrderResponse.java
@@ -2,7 +2,6 @@ package com.foodease.myapp.service.dto.response;
 
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,7 +12,7 @@ import java.util.List;
 public class OrderResponse {
     private Long id;
     private Long userId;
-    private BigDecimal totalPrice;
+    private Long totalPrice;
     private LocalDateTime createdAt;
     private String activeStatus;
     private List<Item> items;
@@ -25,6 +24,6 @@ public class OrderResponse {
     public static class Item {
         private Long id, menuItemId;
         private Integer quantity;
-        private BigDecimal price;
+        private Long price;
     }
 }
